requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
lxml==4.3.2
matplotlib==3.0.3
numpy==1.16.2
Pillow==5.4.1
pkg-resources==0.0.0
pyparsing==2.3.1
python-dateutil==2.8.0
python-pptx==0.6.17
six==1.12.0
XlsxWriter==1.1.5
